Move CollapsibleGroup to collapsible-group/, deprecate expand/ alias
The component lived in the misnamed src/expand/ directory. Duplicate it to a
canonically-named src/collapsible-group/ and keep src/expand/ as a deprecated
alias to be removed in Ring UI 8.0.

- src/collapsible-group/: canonical CollapsibleGroup (component, css, test,
  stories, figma) -- verbatim copy; same directory depth so no import edits.
- src/expand/collapsible-group.tsx: thin re-export shim with @deprecated JSDoc
  and a util-deprecate runtime warning pointing to the new path.
- src/expand/collapsible-group.css: forwards class tokens via `composes` so the
  old CSS-module import path keeps returning the full token map until 8.0.
- src/expand/collapsible-group.test.tsx: compat smoke test for the deprecated
  import (ref/prop forwarding + toggle).
- CHANGELOG.md: document the new path and the deprecation.
- .gitignore: ignore .air/.

The Figma .figma.tsx mapping is left as-is: it never published in either
location (the repo's Code Connect only ingests .figma.js), so proper Code
Connect wiring is tracked separately.
