+++ src/collapsible-group/collapsible-group.figma.css
