+++ src/collapsible-group/collapsible-group.figma.tsx
