+++ src/collapsible-group/collapsible-group.stories.css
